generate viz from context menu
